requirements.txt
pandas>=2.0.0
plotly>=5.18.0
streamlit>=1.29.0
openpyxl>=3.1.0
gspread>=5.12.0
oauth2client>=4.1.3
anthropic>=0.18.0
python-dotenv>=1.0.0
